requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
aiohttp=3.7.4=py39h27cfd23_1
alembic=1.6.5=pyhd8ed1ab_0
async-timeout=3.0.1=py39h06a4308_0
attrs=21.2.0=pyhd3eb1b0_0
blinker=1.4=py_1
brotlipy=0.7.0=py39h3811e60_1001
ca-certificates=2021.5.30=ha878542_0
certifi=2021.5.30=py39hf3d152e_0
cffi=1.14.6=py39he32792d_0
chardet=3.0.4=py39h06a4308_1003
charset-normalizer=2.0.0=pyhd8ed1ab_0
click=8.0.1=pyhd3eb1b0_0
cryptography=3.4.7=py39hbca0aa6_0
dnspython=1.16.0=py39h06a4308_0
dominate=2.6.0=pyhd8ed1ab_0
elasticsearch=7.13.3=pyhd3eb1b0_0
email-validator=1.1.3=pyhd3eb1b0_0
flask=1.1.2=pyhd3eb1b0_0
flask-bootstrap=3.3.7.1=py_0
flask-login=0.5.0=py_0
flask-mail=0.9.1=py_2
flask-migrate=3.0.1=pyhd8ed1ab_1
flask-moment=1.0.2=pyhd8ed1ab_0
flask-sqlalchemy=2.5.1=pyhd3eb1b0_0
flask-wtf=0.15.1=pyhd3eb1b0_0
greenlet=1.1.0=py39h2531618_0
idna=3.2=pyhd3eb1b0_0
importlib-metadata=3.10.0=py39h06a4308_0
itsdangerous=2.0.1=pyhd3eb1b0_0
jinja2=3.0.1=pyhd3eb1b0_0
langdetect=1.0.9=pyhd8ed1ab_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libstdcxx-ng=9.3.0=hd4cf53a_17
mako=1.1.4=pyh44b312d_0
markupsafe=2.0.1=py39h27cfd23_0
multidict=5.1.0=py39h27cfd23_2
ncurses=6.2=he6710b0_1
openssl=1.1.1k=h7f98852_0
pip=21.1.3=py39h06a4308_0
pycparser=2.20=pyh9f0ad1d_2
pyjwt=2.1.0=py39h06a4308_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pysocks=1.7.1=py39hf3d152e_3
python=3.9.5=h12debd9_4
python-dateutil=2.8.1=py_0
python-dotenv=0.18.0=pyhd8ed1ab_0
python-editor=1.0.4=py_0
python_abi=3.9=2_cp39
readline=8.1=h27cfd23_0
redis-py=3.5.3=pyh9f0ad1d_0
requests=2.26.0=pyhd8ed1ab_0
rq=1.9.0=pyhd8ed1ab_0
setuptools=52.0.0=py39h06a4308_0
six=1.16.0=pyh6c4a22f_0
sqlalchemy=1.4.19=py39h7f8727e_0
sqlite=3.36.0=hc218d9a_0
tk=8.6.10=hbc83047_0
typing-extensions=3.10.0.0=hd3eb1b0_0
typing_extensions=3.10.0.0=pyh06a4308_0
tzdata=2021a=h52ac0ba_0
urllib3=1.26.6=pyhd8ed1ab_0
visitor=0.1.3=pyh9f0ad1d_1
werkzeug=1.0.1=pyhd3eb1b0_0
wheel=0.36.2=pyhd3eb1b0_0
wtforms=2.3.3=py_0
xz=5.2.5=h7b6447c_0
yarl=1.6.3=py39h27cfd23_0
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
